fix for timer and some changes for screen resolution

+++ Assets/Scripts/GameManager.cs
@@ -27,7 +27,7 @@ public class GameManager : MonoBehaviour
             time = Time.time - startTime;
 
             string minutes = ((int)time / 60).ToString("00");
-            string seconds = (time % 60).ToString("00");
+            string seconds = ((int)time % 60).ToString("00");
             string milliseconds = ((int)((time * 1000) % 1000)).ToString("000");
 
             timerText.text = minutes + ":" + seconds + ":" + milliseconds;
